+++ test/components/datagrid/datagrid.e2e-spec.js
@@ -2040,6 +2040,22 @@ describe('Datagrid paging with empty dataset', () => {
     await element(by.css('.pager-toolbar .pager-next')).click();
     await browser.driver.sleep(config.sleep);
 
-    expect(await element.all(by.css('#datagrid tbody tr[aria-rowindex]')).count()).toEqual(1);
+    expect(await element.all(by.css('#datagrid tbody tr:nth-child(1)')).count()).toEqual(1);
+  });
+});
+
+describe('Datagrid multiselect sorting test', () => {
+  beforeEach(async () => {
+    await utils.setPage('/components/datagrid/example-grouping-multiselect?layout=nofrills');
+
+    const datagridEl = await element(by.css('#datagrid thead th:nth-child(2)'));
+    await browser.driver
+      .wait(protractor.ExpectedConditions.presenceOf(datagridEl), config.waitsFor);
+  });
+
+  it('Should not have errors', async () => {
+    const thEl = await element(by.css('#datagrid thead th:nth-child(2)'));
+    await thEl.click();
+    await utils.checkForErrors();
   });
 });